teacher.java:add class Dto

+++ src/main/java/com/project/app/dto/TeacherDTO.java
@@ -3,7 +3,7 @@ package com.project.app.dto;
 
 public class TeacherDTO extends HumanDTO {
 
-    public TeacherDTO(int id, int age, String sex, double salary , String type) {
+    public TeacherDTO(Long id, int age, String sex, double salary , String type) {
         super(id,age,sex,salary, type);
     }
 }
